bug 11463: userVms and routers being stopped by the simulator status 11463: resolved fixed reviewed-by: Kishan.Kavala@citrix.com
When the management server restarts and does a full vmSync we need to send back
the state of the simulator VMs in the StartCommand for management server to
restore the states. Or HA will mark them as stopped.
(cherry picked from commit 65526eee33fbd98a9943b9dfce7a1a0c7577caeb)